+++ src/dashboardWebView/constants/SortOption.ts
@@ -2,5 +2,11 @@ export enum SortOption {
   LastModifiedAsc = "LastModifiedAsc",
   LastModifiedDesc = "LastModifiedDesc",
   FileNameAsc = "FileNameAsc",
-  FileNameDesc = "FileNameDesc"
+  FileNameDesc = "FileNameDesc",
+  SizeAsc = "SizeAsc",
+  SizeDesc = "SizeDesc",
+  CaptionAsc = "CaptionAsc",
+  CaptionDesc = "CaptionDesc",
+  AltAsc = "AltAsc",
+  AltDesc = "AltDesc",
 }